+++ src/cli/DB2BaseHandler.ts
@@ -32,7 +32,7 @@ export abstract class DB2BaseHandler implements ICommandHandler {
      */
     public async process(commandParameters: IHandlerParameters) {
         this.mHandlerParams = commandParameters;
-        const session = await DB2Session.createSessCfgFromArgs(commandParameters.arguments);
+        const session = await DB2Session.createSessCfgFromArgs(commandParameters.arguments, true, commandParameters);
         await this.processWithDB2Session(commandParameters, session);
     }