+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to the IBM® Db2® Plug-in for Zowe CLI will be documented in this file.
 
+## Recent Changes
+
+- BugFix: Fixed error messages when host, port, user, password, or database are omitted
+- BugFix: Fixed daemon mode prompting
+
 ## `5.0.0-next.202203241624`
 
 - BugFix: Updated follow-redirects and minimist dependencies to resolve potential vulnerabilities.